refactor package to make it more like spark-avro and spark-csv
